+++ srcjs/components/storage.js
@@ -1,5 +1,6 @@
 import 'shiny'
 import { getStorage, ref, uploadString } from "firebase/storage";
+import { setInputValue } from '../utils.js';
 
 let storage;
 
@@ -19,10 +20,21 @@ export const handleStorage = (firebaseApp) => {
   Shiny.addCustomMessageHandler('fireblaze-upload-file', function(msg) {
 		uploadString(storageRef, msg.encoded, 'data_url')
 			.then((snapshot) => {
-				console.log('Uploaded a base64url string!');
+				if(!msg.response)
+					return;
+
+				let data = {
+					response: snapshot.metadata,
+					success: true
+				}
+
+				setInputValue(msg.response, data, msg.ns);
 			})
 			.catch(error => {
-				console.error(error);
+				if(!msg.response)
+					return;
+				
+				setInputValue(msg.response, {success: false, response: error}, msg.ns);
 			});
 	});
 }